Split ItemLoader tests
